refactor: unify provider priority list into @opencode-ai/util/provider and remove tatu-maas discount tag from provider dialog
