+++ packages/app/src/hooks/use-providers.ts
@@ -1,42 +1,11 @@
 import { useGlobalSync } from "@/context/global-sync"
 import { decode64 } from "@/utils/base64"
+import { popularProviders, rank } from "@opencode-ai/util/provider"
 import { useParams } from "@solidjs/router"
 import { createMemo } from "solid-js"
 
-export const popularProviders = [
-  "tatu-maas",
-  "xiaomi",
-  "xiaomi-token-plan-cn",
-  "alibaba-cn",
-  "alibaba-coding-plan-cn",
-  "deepseek",
-  "moonshotai-cn",
-  "moonshot-cn",
-  "zhipuai",
-  "zhipuai-coding-plan",
-  "minimax-cn",
-  "minimax-cn-coding-plan",
-  "tencent-coding-plan",
-  "siliconflow-cn",
-  "baidu",
-  "qianfan",
-  "ernie",
-  "baidu-qianfan",
-  "opencode",
-  "opencode-go",
-  "anthropic",
-  "openai",
-  "github-copilot",
-  "google",
-  "openrouter",
-  "vercel",
-]
+export { popularProviders, rank }
 const popularProviderSet = new Set(popularProviders)
-
-export function rank(id: string) {
-  const index = popularProviders.indexOf(id)
-  return index >= 0 ? index : popularProviders.length
-}
 
 export function useProviders() {
   const globalSync = useGlobalSync()
