+++ packages/opencode/src/cli/cmd/tui/component/provider-priority.ts
@@ -1,33 +1,3 @@
-const ranked = [
-  "tatu-maas",
-  "xiaomi",
-  "xiaomi-token-plan-cn",
-  "alibaba-cn",
-  "alibaba-coding-plan-cn",
-  "deepseek",
-  "moonshotai-cn",
-  "moonshot-cn",
-  "zhipuai",
-  "zhipuai-coding-plan",
-  "minimax-cn",
-  "minimax-cn-coding-plan",
-  "tencent-coding-plan",
-  "siliconflow-cn",
-  "baidu",
-  "qianfan",
-  "ernie",
-  "baidu-qianfan",
-  "opencode",
-  "opencode-go",
-  "anthropic",
-  "openai",
-  "github-copilot",
-  "google",
-  "openrouter",
-  "vercel",
-]
+import { rank } from "@opencode-ai/util/provider"
 
-export function rank(id: string) {
-  const index = ranked.indexOf(id)
-  return index >= 0 ? index : ranked.length
-}
+export { rank }
